arm: update MMU for ARMv7 and enable client mode.
Previously we were using manager mode which was hiding a number of
issues as no fault checking is performed in that mode.  The various
access bits are now mostly correct, though the L2 entries have
read-write permissions for now just to get things booting.  I don't know
if that's a bug or a misunderstanding of the documentation on my part.