ARM Allow a basic 'make setup' for ARM.
This fills in sufficient stub-y parts of the build that one can complete
an ARM 'make setup', albiet with some parts deliberately excised, for
the present. This includes the tools build as well. Parts of this were
done by Rich Lowe in the past. A lot of it was adapted for the modern
(2019) system.